requirements.txt
# Core web backend
flask
flask-cors

# ZIM file reading
libzim
beautifulsoup4

# LLM / RAG stack
langchain>=0.2
langchain-community
faiss-cpu      # vector store
chromadb       # optional alternative to FAISS
tiktoken       # for token counting (optional)

# Utilities
tqdm